### 2021-08-20 #3639 @whedon
[PRE REVIEW]: Heta compiler: a software tool for the development of large-scale QSP models and compilation into simulation formats
**Submitting author:** @metelkin (<a href="http://orcid.org/0000-0001-6612-5373">Evgeny Metelkin</a>)
**Repository:** <a href="https://github.com/hetalang/heta-compiler" target ="_blank">https://github.com/hetalang/heta-compiler</a>
**Version:** v0.6.6
