+++ README.md
@@ -3,7 +3,8 @@
 [![Autotests](https://github.com/hetalang/heta-compiler/workflows/Autotests/badge.svg)](https://github.com/hetalang/heta-compiler/actions)
 [![Coverage Status](https://coveralls.io/repos/github/hetalang/heta-compiler/badge.svg?branch=master)](https://coveralls.io/github/hetalang/heta-compiler?branch=master)
 [![GitHub npm](https://img.shields.io/npm/v/heta-compiler/latest.svg)](https://www.npmjs.com/package/heta-compiler)
-[![api-docs](https://img.shields.io/website?down_color=yellow&label=Docs&up_color=green&url=https%3A%2F%2Fhetalang.github.io%2F#%2Fheta-compiler%2F)](https://hetalang.github.io/#/heta-compiler/)
+[![api-docs](https://img.shields.io/website?down_color=yellow&label=api-docs&up_color=green&url=https%3A%2F%2Fhetalang.github.io%2Fheta-compiler%2Fdev%2F)](https://hetalang.github.io/heta-compiler/dev/)
+[![status](https://joss.theoj.org/papers/ebff76c368d3adb720afe414ef6b29fb/status.svg)](https://joss.theoj.org/papers/ebff76c368d3adb720afe414ef6b29fb)
 [![GitHub license](https://img.shields.io/github/license/hetalang/heta-compiler.svg)](https://github.com/hetalang/heta-compiler/blob/master/LICENSE)
 
 # Heta compiler
@@ -116,7 +117,7 @@ The following is the example where we create a Heta module and compile it into S
 
 Heta compiler has been written in NodeJS environment and can be used as a package for browser or server-side tools and applications.
 
-> To learn more more, see [API references](https://hetalang.github.io/heta-compiler/dev)
+> To learn more more, see [API docs](https://hetalang.github.io/heta-compiler/dev)
 (under development).
 
 ```javascript
