### 2021-08-23 #3639 @danielskatz
@metelkin - thanks for your submission.  Because the code size is not clear, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.

Can you help by explaining specifically what code is part of the submission that would be reviewed?
