### 2021-08-23 #3639 @danielskatz
I guess I should have read https://hetalang.github.io/#/heta-compiler/ - now I see that the Heta compiler has been written in NodeJS environment, which I think answer my question about the source code. And given this, I think my question about code size is also answered.  So I'll stop the scope query, and work on finding an editor next.
