### 2021-08-25 #3639 @metelkin
Hi @fboehm. Thank you for your editing.
I look through the list of people who have already agreed to review for JOSS. I guess potential reviewers might be

sahilseth
louiejtaylor
camillescott
zfrenchee
RuneBlaze
jmcgonigle
smsaladi
dvanic
carleshf
jkanche
