### 2021-08-31 #3639 @fboehm
:wave: @lgerosa and @bence-szalai - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks for considering this.
