### 2021-09-01 f65f923
API docs for `Builder`
