+++ api-references.md
@@ -1,20 +1,21 @@
 # API references
 
-*Heta parser API references page is under developments. If you have questions contact the developers directly.*
+This is a documentation for the JavaScript interface of __Heta compiler__.
+
+
+*Under developments. If you have questions contact the developers directly.*
 
 ## Classes
 
-- Builder
-- Container
-- Namespace
-- Logger
-- Unit
-- Expression
-- ModuleSystem
-- _Module
+- [Builder]{@link Builder}
+- [Container]{@link Container}
+- [Namespace]{@link Namespace}
+- [Logger]{@link Logger}
+- [Unit]{@link Unit}
+- [Expression]{@link Expression}
+- [ModuleSystem]{@link ModuleSystem}
+- [_Module]{@link _Module}
 
 - Component
 - _Size
 - Const
-
-...
