### 2021-09-01 #3639 @fboehm
:wave: @jmcgonigle, @smsaladi, @dvanic, @carleshf, @jkanche - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks for considering this.
