### 2021-09-02 561a64b
`Component` API docs, part 1
