+++ src/container/main.js
@@ -23,7 +23,27 @@ const { Logger, JSONTransport } = require('../logger');
 const coreItems = require('./core-items.json');
 const TopoSort = require('@insysbio/topo-sort');
 
-
+/**
+ * The main class storing a modeling platform and it's methods.
+ * 
+ * It is highly recommended to use only one container instance in developed code.
+ * 
+ * @class Container
+ * 
+ * @property {object} classes Map-like storage for all element constructors that can be created inside platform.
+ *    For example the element of the type `UnitsDef` can be created as follows:
+ *    ```let new_unit = new c.UnitDef({id: 'new', units: 'g/litre'})```
+ *    The `new_unit` element will be automatically bound to the container and pushed to `unitDefStorage`.
+ * @property {Logger} logger object providing transport of errors, warnings and info messages on Heta platform level.
+ * @property {object[]} defaultLogs Default storage of errors which will be used for diagnostics.
+ *    The {@link JSONTransport} is used here.
+ * @property {Map<string,_Export>} exportStorage Storage for `_Export` instances. Key is a string identifier.
+ * @property {Map<string,UnitDef>} unitDefStorage Storage for `UnitDef` instances. Key is a string identifier.
+ * @property {Map<string,Namespace>} namespaceStorage Storage for `Namespace` instances. Key is a string identifier.
+ *    There is a default namespace with identifier `nameless` which will be used as a default namespace 
+ *    for all components where namespace name is not set.
+ * @property {object} _componentClasses map-like structure for storing all available classes describing `Component`s.
+ */
 class Container {
   /* constructor can be run many times */
   constructor(){
@@ -67,11 +87,31 @@ class Container {
     this.loadMany(coreItems, true);
     //console.log(this.defaultLogs)
   }
-  // returns array of errors in heta code
+
+  /**
+   * Returns array of errors from the default logger.
+   * 
+   * @method Container#hetaErrors
+   * 
+   * @returns {object[]} See details in {@link JSONTransport}
+   */
   hetaErrors(){
     return this.defaultLogs
       .filter(x => x.levelNum >= 3);
   }
+
+  /**
+   * Runs an action (like creating a component) based on `q.action` property.
+   * If `q.action` is not set than apply "upsert".
+   * An "action" name should be set as a name of the `Container` method. 
+   * 
+   * This is the main method to convert from elements of "queue" into platform elements.
+   * 
+   * @param {object} q Simple object with the same structure as Heta plain format.
+   * @param {boolean} isCore Set element as a "core" which means you cannot rewrite or delete it.
+   * 
+   * @returns {Container} This function returns the container.
+   */
   load(q, isCore = false){
     // estimate action, default is upsert
     let actionName = _.get(q, 'action', 'upsert');
@@ -85,17 +125,38 @@ class Container {
     // normal flow
     return this[actionName](q, isCore);
   }
+
+  /**
+   * Runs {@link Container#load} method many times for each element of `qArr` vector sequentially.
+   * 
+   * @param {object[]} qArr Array of "queue" formatted elements.
+   * @param {boolean} isCore Set element as a "core" which means you cannot rewrite or delete it.
+   * 
+   * @returns {Container} This function returns the container.
+   */
   loadMany(qArr, isCore = false){
     qArr.forEach((q) => this.load(q, isCore));
     return this;
   }
+
+  /**
+   * Get number of total elements of a platform.
+   * 
+   * @returns {number} Total number of components + `UnitDef` + `_Export` instances.
+   */
   get length(){
     return [...this.namespaceStorage]
       .reduce((acc, x) => acc + x[1].size, 0)
[…]
