+++ api-references.md
@@ -2,6 +2,7 @@
 
 This is a documentation for the JavaScript interface of __Heta compiler__.
 
+For users guidance and CLI references see the main [documentation site](https://hetalang.github.io/#/heta-compiler/).
 
 *Under developments. If you have questions contact the developers directly.*
 
