### 2021-09-04 #3639 @fboehm
@lptolik and @Ahmed7emdan - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks for considering this.
