### 2021-09-05 68fe08d
api docs for namespace
