+++ src/namespace.js
@@ -2,27 +2,51 @@ const TopoSort = require('@insysbio/topo-sort');
 const _ = require('lodash');
 const { flatten } = require('./core/utilities');
 
-// implicit property container, XXX: should be renamed to _container
 class Namespace extends Map {
+  /**
+   * Namespace is a `Map` which stores components.
+   * 
+   * @extends Map
+   * 
+   * @param {string} spaceName Space identifier.
+   */
   constructor(spaceName){ 
     super();
     //if (typeof spaceName !== 'string')
     //  throw new TypeError(`spaceName argument must be string, got ${spaceName}`);
     this._spaceName = spaceName;
   }
+  /**
+   * `true` if namespace is abstract, `false` if concrete.
+   */
   get isAbstract(){
     return this._isAbstract;
   }
-  // title of space
+  /**
+   * Identifier of a namespace.
+   */
   get spaceName(){
     return this._spaceName;
   }
   set(key, value){
     return super.set(key, value);
   }
+  /**
+   * Converts namespace into array of components.
+   * 
+   * @returns {Component[]} All components in array format.
+   */
   toArray(){
     return [...this].map((x) => x[1]);
   }
+  /**
+   * Converts namespace into queue array.
+   * 
+   * @param {boolean} removeCoreComponents=true 
+   * @param {object} options Options passed to {@link Top#toQ} method
+   * 
+   * @returns {object[]} Array of JS objects in queue format.
+   */
   toQArr(removeCoreComponents = false, options = {}){
     let qArr = this.toArray()
       .filter((x) => !(removeCoreComponents && x.isCore))
@@ -30,6 +54,12 @@ class Namespace extends Map {
     
     return qArr;
   }
+  /**
+   * Representation of namespace in queue format.
+   * 
+   * @param {object} options reserved for future versions.
+   * @returns {object} JS object in queue format.
+   */
   toQ(options = {}){
     let res = {
       action: 'setNS',
@@ -39,24 +69,55 @@ class Namespace extends Map {
 
     return res; 
   }
+
+  /**
+   * Representation of namespace in flat queue format.
+   * 
+   * @param {object} options reserved for future versions.
+   * 
+   * @returns {object} JS object in flat queue format.
+   */
   toFlat(options = {}){
     let q = this.toQ(options);
     let res = flatten(q);
 
     return res; 
   }
+
+  /**
+   * Components of the particular class.
+   * 
+   * @param {string} className One of available Heta classes: 'Const', 'Record', etc.
+   * 
+   * @returns {Component[]} Array of selected components.
+   */
   selectByClassName(className){
     return this
       .toArray()
       .filter((x) => x.className === className);
   }
+
+  /**
+   * Components of the class and its inheritors.
+   * 
+   * @param {string} className One of available Heta classes: 'Const', 'Record', etc.
+   * 
+   * @returns {Component[]} Array of selected components.
+   */
   selectByInstanceOf(className){
     return this
       .toArray()
       .filter((x) => x.instanceOf(className));
   }
-  /*
-  */
+
+  /**
+   * Select all records from namespace and sort them by the scope.
+   * 
+   * @param {string} context scope of `Record`.
+   * @param {boolean} includeCompartmentDep=true
+   * 
+   * @returns {Component[]} Records in scope order.
+   */
   sortExpressionsByContext(context, includeCompartmentDep = false){
     // create topo-sort tree
     let graph = new TopoSort();
@@ -98,10 +159,25 @@ class Namespace extends Map {
 
     return sorted;
   }
+
+  /**
+   * 
+   * Select all records which have the scope.
+   * 
+   * @param {string} context scope of `Record`.
+   * 
[…]
