### 2021-09-07 #3639 @martinmodrak
Hi, the application area is mostly outside my expertise, but I can review the software engineering aspects of the work, if that's enough.
