### 2021-09-09 4aca352
API docs for module types
