+++ src/module-system/heta-module.js
@@ -2,6 +2,13 @@ const fs = require('fs');
 const hetaParser = require('heta-parser');
 const _Module = require('./module');
 
+/**
+ * To initialize a Heta module of the "heta" type.
+ * It includes reading and parsing file formatted as Heta code,
+ * see [Heta specifications](https://hetalang.github.io/#/specifications/modules?id=heta-module)
+ * 
+ * @returns {Module} Self.
+ */
 _Module.prototype.setHetaModule = function(){
   try {
     let fileContent = fs.readFileSync(this.filename, 'utf8');
@@ -15,11 +22,17 @@ _Module.prototype.setHetaModule = function(){
   return this;
 };
 
-// wrapper for SyntaxError to show the coordinates
+/**
+ * This method is a wrapper for `parse()` method of "heta-parser" package to show errors location.
+ * 
+ * @param {string} filename File to parse. It is used only for log messages.
+ * @param  {...any} params Additional parameters passed to `parse()` method.
+ * @returns {array} Queue array format.
+ */
 function _hetaParse(filename, ...params){
-  try{
+  try {
     return hetaParser.parse(...params);
-  }catch(e){
+  } catch(e) {
     if((e instanceof hetaParser.SyntaxError)){
       e.name = 'HetaParsingError';
       let loc = e.location;
