+++ src/module-system/json-module.js
@@ -1,6 +1,13 @@
 const fs = require('fs');
 const _Module = require('./module');
 
+/**
+ * To initialize a Heta module of the "json" type.
+ * It includes reading and parsing file formatted as Heta-JSON,
+ * see [Heta specifications](https://hetalang.github.io/#/specifications/modules?id=json-module)
+ * 
+ * @returns {Module} Self.
+ */
 _Module.prototype.setJSONModule = function(){
   //checking file exists
   let fileContent = fs.readFileSync(this.filename, 'utf8');
@@ -15,16 +22,25 @@ _Module.prototype.setJSONModule = function(){
   return this;
 };
 
+/**
+ * Auxillary function to parse JSON file in {@link _Module#setJSONModule} method.
+ * This function is a wrapper of `JSON.parse()` method to get more clear description of errors.
+ * 
+ * @param {string} filename File to parse. It is used only for log messages.
+ * @param  {...any} params Additional parameters passed to `JSON.parse()` method.
+ * 
+ * @returns {array} An object representing JSON file content.
+ */
 function _JSONParse(
   filename, // used here to display in error message 
   ...params
 ){
-  try{
+  try {
     return JSON.parse(...params);
-  }catch(e){
+  } catch(e) {
     let select = e.message.match(/at position (\d*)/); // This is ugly part, sorry
     if((e instanceof SyntaxError) && typeof +select[1]==='number'){
-      e.name = 'JSONSynaxError';
+      e.name = 'JSONSyntaxError';
       let pos = +select[1];
       let parsedPart = params[0].substring(0, pos);
       let splittedText = parsedPart.split(/\r*\n/);
