+++ src/module-system/sbml-module.js
@@ -4,7 +4,7 @@ const { SBMLParse } = require('./sbml-parse');
 
 /**
  * To initialize a Heta module of the "sbml" type.
- * It includes reading and parsing SBML file and representing it into Heta queue,
+ * It includes reading and parsing SBML file and representing it into Q-object,
  * see [Heta specifications](https://hetalang.github.io/#/specifications/modules?id=sbml-module)
  * 
  * @returns {_Module} Self.
