### 2021-09-16 f4f4cce
QueueError => QError
