+++ src/container/actions.js
@@ -26,14 +26,14 @@ Container.prototype.export = function(q = {}, isCore = false){
   if (q.format === undefined) {
     this.logger.error(
       'Empty "format" option in #export',
-      {type: 'QueueError'}
+      {type: 'QError'}
     );
     return;
   }
   if (typeof this.classes[q.format] !== 'function') {
     this.logger.error(
       `Unknown format "${q.format}" in #export action.`,
-      {type: 'QueueError'}
+      {type: 'QError'}
     );
     return;
   }
@@ -80,21 +80,21 @@ Container.prototype.insert = function(q = {}, isCore = false){
   if (!q.id || !/^[_a-zA-Z][_a-zA-Z0-9]*$/.test(q.id)) {
     this.logger.error(
       `${ind} id should be string of type ID, but have "${q.id}"`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
   if (reservedWords.indexOf(q.id) !== -1) {
     this.logger.error(
       `${ind} id cannot be one of reserved word, but have "${q.id}". reservedWords = [${reservedWords}]`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
   if (!q.class || typeof q.class !== 'string'){
     this.logger.error(
       `${ind} No class or unsuitable class for "insert": ${q.class}`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
@@ -116,7 +116,7 @@ Container.prototype.insert = function(q = {}, isCore = false){
   if (selectedClass === undefined){
     this.logger.error(
       `"${ind}" Unknown class "${q.class}" for the component.`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
@@ -126,7 +126,7 @@ Container.prototype.insert = function(q = {}, isCore = false){
   if (namespace === undefined) {
     this.logger.error(
       `"${ind}" create namespace "${space}" before use.`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
@@ -154,14 +154,14 @@ Container.prototype.update = function(q = {}){
   if (!q.id || !/^[_a-zA-Z][_a-zA-Z0-9]*$/.test(q.id)) {
     this.logger.error(
       `${ind} Id should be string, but have "${q.id}"`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
   if (q.class){
     this.logger.error(
       `${ind} Class property is not allowed for "update": ${q.class}`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
@@ -170,7 +170,7 @@ Container.prototype.update = function(q = {}){
   if (namespace === undefined) {
     this.logger.error(
       `${ind} Create namespace "${space}" before use.`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
@@ -180,14 +180,14 @@ Container.prototype.update = function(q = {}){
   if (targetComponent === undefined) {
     this.logger.error(
       `${ind} component does not exist which is not allowed for the "update" action.`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
   if (targetComponent.isCore) {
     this.logger.error(
       `${ind} Core component is read-only and cannot be updated.`,
-      {type: 'QueueError', space: space}
+      {type: 'QError', space: space}
     );
     return;
   }
[…]
