### 2021-10-07 05ad142
formal support for setScenario actions
