+++ src/xlsx-export/index.js
@@ -43,7 +43,7 @@ class XLSXExport extends AbstractExport {
     return 'XLSXExport';
   }
   get format(){
-    return 'XLSX'
+    return 'XLSX';
   }
   make(){
     // filtered namespaces
@@ -62,7 +62,10 @@ class XLSXExport extends AbstractExport {
     let fArr_unitDef = [...this._container.unitDefStorage]
       .filter((x) => !x[1].isCore)
       .map((x) => x[1].toFlat());
-    let fArr_full = [].concat(fArr_ns, fArr_unitDef).map((x) => {
+    let fArr_scenario = [...this._container.scenarioStorage]
+      .filter((x) => !x[1].isCore)
+      .map((x) => x[1].toFlat());
+    let fArr_full = [].concat(fArr_ns, fArr_unitDef, fArr_scenario).map((x) => {
       x.on = 1;
       return _.mapValues(x, (value) => typeof value === 'boolean' ? value.toString() : value);
     });
