+++ src/yaml-export/index.js
@@ -47,7 +47,11 @@ class YAMLExport extends AbstractExport {
     let qArr_unitDef = [...this._container.unitDefStorage]
       .filter((x) => !x[1].isCore)
       .map((x) => x[1].toQ());
-    let qArr_full = [].concat(qArr_ns, qArr_unitDef);
+    let qArr_scenario = [...this._container.scenarioStorage]
+      .filter((x) => !x[1].isCore)
+      .map((x) => x[1].toQ());
+      
+    let qArr_full = [].concat(qArr_ns, qArr_unitDef, qArr_scenario);
 
     // remove unnecessary properties
     let qArr = this.omit ? qArr_full.map((q) => _.omit(q, this.omit)) : qArr_full;
