+++ export-formats.md
@@ -368,4 +368,4 @@ Creation of Matlab files (.m) which represent ODE and code to run ODE.
 |MathExpr: `e`, `pi`                                     |+ |+ |+ |+ |+ |+ |+ |+ |+
 |MathExpr: `Infinity`, `NaN`                             |- |- |+ |+ |+ |+ |+ |+ |+
 |Const: `Infinity`, `NaN`                                |- |- |+ |+ |+ |+ |+ |+ |+
-|`@Scenario` support                                     |- |- |- |- |- |- |- |- |-
+|`@Scenario` support                                     |- |- |- |- |- |- |- |+ |+
