### 2021-11-01 hetalang/heta-compiler#10 @martinmodrak
The jak2stat5 example fails with mrgsolve
I took the model at https://github.com/hetalang/qsp-benchmark/tree/master/jak2stat5/src, compiled it and tried to run it with `mrgsolve`. Unfortunately, I get an error.

I was able to run the simple example model from readme via mrgsolve, so it presumably isn't just an installation issue.

The error I get is:

```
> nameless_model <- mrgsolve::mread(model = 'nameless', file = 'model.cpp')
Building nameless ... error.

---:: stderr ::---------------------------------------------
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:25:10: error: expected unqualified-id before 'default'
   double default;
          ^~~~~~~

7:1: error: case label not within a switch statement
 default = 1.0;
 ^~~~~~~
7:8: error: expected ':' before '=' token
 default = 1.0;
        ^~
        :
7:9: error: expected primary-expression before '=' token
 default = 1.0;
         ^

14:37: error: expected primary-expression before 'default'
 EpoRJAK2_amt__0 = (init_EpoRJAK2) * default;
                                     ^~~~~~~
15:24: error: expected primary-expression before 'default'
 SOCS3_amt__0 = (0.0) * default;
                        ^~~~~~~
16:28: error: expected primary-expression before 'default'
 EpoRpJAK2_amt__0 = (0.0) * default;
                            ^~~~~~~
17:26: error: expected primary-expression before 'default'
 SHP1Act_amt__0 = (0.0) * default;
                          ^~~~~~~
18:31: error: expected primary-expression before 'default'
 EpoRJAK2_CIS_amt__0 = (0.0) * default;
                               ^~~~~~~
19:30: error: expected primary-expression before 'default'
 p1EpoRpJAK2_amt__0 = (0.0) * default;
                              ^~~~~~~
20:30: error: expected primary-expression before 'default'
 p2EpoRpJAK2_amt__0 = (0.0) * default;
                              ^~~~~~~
21:31: error: expected primary-expression before 'default'
 p12EpoRpJAK2_amt__0 = (0.0) * default;
                               ^~~~~~~
22:29: error: expected primary-expression before 'default'
 SHP1_amt__0 = (init_SHP1) * default;
                             ^~~~~~~
24:22: error: expected primary-expression before 'default'
 CIS_amt__0 = (0.0) * default;
                      ^~~~~~~
In file included from nameless-mread-source.cpp:3:

nameless-mread-header.h:89:35: error: expected primary-expression before 'default'
 #define EpoRJAK2 (EpoRJAK2_amt_ / default)
                                   ^~~~~~~
44:13: note: in expansion of macro 'EpoRJAK2'
 v1 = (Epo * EpoRJAK2 * JAK2ActEpo) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
             ^~~~~~~~
nameless-mread-header.h:89:35: error: expected ')' before 'default'
 #define EpoRJAK2 (EpoRJAK2_amt_ / default)
                  ~                ^~~~~~~
44:13: note: in expansion of macro 'EpoRJAK2'
 v1 = (Epo * EpoRJAK2 * JAK2ActEpo) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
             ^~~~~~~~
44:75: error: expected ')' before ';' token
 v1 = (Epo * EpoRJAK2 * JAK2ActEpo) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
      ~                                                                    ^
                                                                           )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:90:37: error: expected primary-expression before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                                     ^~~~~~~
45:6: note: in expansion of macro 'EpoRpJAK2'
 v2 = EpoRpJAK2 * (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act;
      ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected ')' before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                   ~                 ^~~~~~~
45:6: note: in expansion of macro 'EpoRpJAK2'
 v2 = EpoRpJAK2 * (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act;
      ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected primary-expression before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                                     ^~~~~~~
46:7: note: in expansion of macro 'EpoRpJAK2'
 v3 = (EpoRpJAK2 * EpoRActJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
       ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected ')' before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                   ~                 ^~~~~~~
46:7: note: in expansion of macro 'EpoRpJAK2'
 v3 = (EpoRpJAK2 * EpoRActJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
       ^~~~~~~~~
46:71: error: expected ')' before ';' token
 v3 = (EpoRpJAK2 * EpoRActJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
      ~                                                                ^
                                                                       )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:90:37: error: expected primary-expression before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                                     ^~~~~~~
47:13: note: in expansion of macro 'EpoRpJAK2'
 v4 = (3.0 * EpoRpJAK2 * EpoRActJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
             ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected ')' before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                   ~                 ^~~~~~~
47:13: note: in expansion of macro 'EpoRpJAK2'
 v4 = (3.0 * EpoRpJAK2 * EpoRActJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
             ^~~~~~~~~
47:115: error: expected ')' before ';' token
 v4 = (3.0 * EpoRpJAK2 * EpoRActJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
      ~                                                                                                            ^
                                                                                                                   )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:91:41: error: expected primary-expression before 'default'
 #define p1EpoRpJAK2 (p1EpoRpJAK2_amt_ / default)
                                         ^~~~~~~
48:27: note: in expansion of macro 'p1EpoRpJAK2'
 v5 = (3.0 * EpoRActJAK2 * p1EpoRpJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
                           ^~~~~~~~~~~
nameless-mread-header.h:91:41: error: expected ')' before 'default'
 #define p1EpoRpJAK2 (p1EpoRpJAK2_amt_ / default)
                     ~                   ^~~~~~~
48:27: note: in expansion of macro 'p1EpoRpJAK2'
 v5 = (3.0 * EpoRActJAK2 * p1EpoRpJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
                           ^~~~~~~~~~~
48:117: error: expected ')' before ';' token
 v5 = (3.0 * EpoRActJAK2 * p1EpoRpJAK2) / ((EpoRCISInh * EpoRJAK2_CIS + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0));
      ~                                                                                                              ^
                                                                                                                     )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:92:41: error: expected primary-expression before 'default'
 #define p2EpoRpJAK2 (p2EpoRpJAK2_amt_ / default)
                                         ^~~~~~~
49:21: note: in expansion of macro 'p2EpoRpJAK2'
 v6 = (EpoRActJAK2 * p2EpoRpJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
                     ^~~~~~~~~~~
nameless-mread-header.h:92:41: error: expected ')' before 'default'
 #define p2EpoRpJAK2 (p2EpoRpJAK2_amt_ / default)
                     ~                   ^~~~~~~
49:21: note: in expansion of macro 'p2EpoRpJAK2'
 v6 = (EpoRActJAK2 * p2EpoRpJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
                     ^~~~~~~~~~~
49:73: error: expected ')' before ';' token
 v6 = (EpoRActJAK2 * p2EpoRpJAK2) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0);
      ~                                                                  ^
                                                                         )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:96:33: error: expected primary-expression before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                                 ^~~~~~~
50:38: note: in expansion of macro 'SHP1Act'
 v7 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p1EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:96:33: error: expected ')' before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                 ~               ^~~~~~~
50:38: note: in expansion of macro 'SHP1Act'
 v7 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p1EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:96:33: error: expected primary-expression before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                                 ^~~~~~~
51:38: note: in expansion of macro 'SHP1Act'
 v8 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p2EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:96:33: error: expected ')' before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                 ~               ^~~~~~~
51:38: note: in expansion of macro 'SHP1Act'
 v8 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p2EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:96:33: error: expected primary-expression before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                                 ^~~~~~~
52:38: note: in expansion of macro 'SHP1Act'
 v9 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p12EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:96:33: error: expected ')' before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                 ~               ^~~~~~~
52:38: note: in expansion of macro 'SHP1Act'
 v9 = (JAK2EpoRDeaSHP1 / init_SHP1) * SHP1Act * p12EpoRpJAK2;
                                      ^~~~~~~
nameless-mread-header.h:94:43: error: expected primary-expression before 'default'
 #define EpoRJAK2_CIS (EpoRJAK2_CIS_amt_ / default)
                                           ^~~~~~~
53:7: note: in expansion of macro 'EpoRJAK2_CIS'
 v10 = EpoRJAK2_CIS * (EpoRCISRemove / init_EpoRJAK2) * (p12EpoRpJAK2 + p1EpoRpJAK2);
       ^~~~~~~~~~~~
nameless-mread-header.h:94:43: error: expected ')' before 'default'
 #define EpoRJAK2_CIS (EpoRJAK2_CIS_amt_ / default)
                      ~                    ^~~~~~~
53:7: note: in expansion of macro 'EpoRJAK2_CIS'
 v10 = EpoRJAK2_CIS * (EpoRCISRemove / init_EpoRJAK2) * (p12EpoRpJAK2 + p1EpoRpJAK2);
       ^~~~~~~~~~~~
nameless-mread-header.h:95:27: error: expected primary-expression before 'default'
 #define SHP1 (SHP1_amt_ / default)
                           ^~~~~~~
54:7: note: in expansion of macro 'SHP1'
 v11 = SHP1 * (SHP1ActEpoR / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2);
       ^~~~
nameless-mread-header.h:95:27: error: expected ')' before 'default'
 #define SHP1 (SHP1_amt_ / default)
              ~            ^~~~~~~
54:7: note: in expansion of macro 'SHP1'
 v11 = SHP1 * (SHP1ActEpoR / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2);
       ^~~~
nameless-mread-header.h:96:33: error: expected primary-expression before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                                 ^~~~~~~
55:17: note: in expansion of macro 'SHP1Act'
 v12 = SHP1Dea * SHP1Act;
                 ^~~~~~~
nameless-mread-header.h:96:33: error: expected ')' before 'default'
 #define SHP1Act (SHP1Act_amt_ / default)
                 ~               ^~~~~~~
55:17: note: in expansion of macro 'SHP1Act'
 v12 = SHP1Dea * SHP1Act;
                 ^~~~~~~
nameless-mread-header.h:90:37: error: expected primary-expression before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                                     ^~~~~~~
56:50: note: in expansion of macro 'EpoRpJAK2'
 v13 = (STAT5 * (STAT5ActJAK2 / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2)) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0) * cyt;
                                                  ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected ')' before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                   ~                 ^~~~~~~
56:50: note: in expansion of macro 'EpoRpJAK2'
 v13 = (STAT5 * (STAT5ActJAK2 / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2)) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0) * cyt;
                                                  ^~~~~~~~~
56:150: error: expected ')' before ';' token
 v13 = (STAT5 * (STAT5ActJAK2 / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2)) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0) * cyt;
                                                 ~                                                                                                    ^
                                                                                                                                                      )
56:150: error: expected ')' before ';' token
 v13 = (STAT5 * (STAT5ActJAK2 / init_EpoRJAK2) * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2)) / (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0) * cyt;
       ~                                                                                                                                              ^
                                                                                                                                                      )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:93:43: error: expected primary-expression before 'default'
 #define p12EpoRpJAK2 (p12EpoRpJAK2_amt_ / default)
                                           ^~~~~~~
57:64: note: in expansion of macro 'p12EpoRpJAK2'
 v14 = (STAT5 * (STAT5ActEpoR / pow(init_EpoRJAK2, 2.0)) * pow((p12EpoRpJAK2 + p1EpoRpJAK2), 2.0)) / ((CIS * (CISInh / CISEqc) + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0)) * cyt;
                                                                ^~~~~~~~~~~~
nameless-mread-header.h:93:43: error: expected ')' before 'default'
 #define p12EpoRpJAK2 (p12EpoRpJAK2_amt_ / default)
                      ~                    ^~~~~~~
57:64: note: in expansion of macro 'p12EpoRpJAK2'
 v14 = (STAT5 * (STAT5ActEpoR / pow(init_EpoRJAK2, 2.0)) * pow((p12EpoRpJAK2 + p1EpoRpJAK2), 2.0)) / ((CIS * (CISInh / CISEqc) + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0)) * cyt;
                                                                ^~~~~~~~~~~~
57:180: error: expected ')' before ';' token
 v14 = (STAT5 * (STAT5ActEpoR / pow(init_EpoRJAK2, 2.0)) * pow((p12EpoRpJAK2 + p1EpoRpJAK2), 2.0)) / ((CIS * (CISInh / CISEqc) + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0)) * cyt;
                                                               ~                                                                                                                    ^
                                                                                                                                                                                    )
57:180: error: expected ')' before ';' token
 v14 = (STAT5 * (STAT5ActEpoR / pow(init_EpoRJAK2, 2.0)) * pow((p12EpoRpJAK2 + p1EpoRpJAK2), 2.0)) / ((CIS * (CISInh / CISEqc) + 1.0) * (SOCS3 * (SOCS3Inh / SOCS3Eqc) + 1.0)) * cyt;
       ~                                                                                                                                                                            ^
                                                                                                                                                                                    )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:106:25: error: expected primary-expression before 'default'
 #define CIS (CIS_amt_ / default)
                         ^~~~~~~
68:7: note: in expansion of macro 'CIS'
 v25 = CIS * CISTurn;
       ^~~
nameless-mread-header.h:106:25: error: expected ')' before 'default'
 #define CIS (CIS_amt_ / default)
             ~           ^~~~~~~
68:7: note: in expansion of macro 'CIS'
 v25 = CIS * CISTurn;
       ^~~
nameless-mread-header.h:113:29: error: expected primary-expression before 'default'
 #define SOCS3 (SOCS3_amt_ / default)
                             ^~~~~~~
78:7: note: in expansion of macro 'SOCS3'
 v35 = SOCS3 * SOCS3Turn;
       ^~~~~
nameless-mread-header.h:113:29: error: expected ')' before 'default'
 #define SOCS3 (SOCS3_amt_ / default)
               ~             ^~~~~~~
78:7: note: in expansion of macro 'SOCS3'
 v35 = SOCS3 * SOCS3Turn;
       ^~~~~
nameless-mread-header.h:90:37: error: expected primary-expression before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                                     ^~~~~~~
81:64: note: in expansion of macro 'EpoRpJAK2'
 pJAK2_au = log10(offset_pJAK2_long + 2.0 * scale_pJAK2_long * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                                ^~~~~~~~~
nameless-mread-header.h:90:37: error: expected ')' before 'default'
 #define EpoRpJAK2 (EpoRpJAK2_amt_ / default)
                   ~                 ^~~~~~~
81:64: note: in expansion of macro 'EpoRpJAK2'
 pJAK2_au = log10(offset_pJAK2_long + 2.0 * scale_pJAK2_long * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                                ^~~~~~~~~
81:134: error: expected ')' before ';' token
 pJAK2_au = log10(offset_pJAK2_long + 2.0 * scale_pJAK2_long * (EpoRpJAK2 + p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                               ~                                                                      ^
                                                                                                                                      )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:93:43: error: expected primary-expression before 'default'
 #define p12EpoRpJAK2 (p12EpoRpJAK2_amt_ / default)
                                           ^~~~~~~
82:65: note: in expansion of macro 'p12EpoRpJAK2'
 pEpoR_au = log10(offset_pEpoR_long + 16.0 * scale_pEpoR_long * (p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                                 ^~~~~~~~~~~~
nameless-mread-header.h:93:43: error: expected ')' before 'default'
 #define p12EpoRpJAK2 (p12EpoRpJAK2_amt_ / default)
                      ~                    ^~~~~~~
82:65: note: in expansion of macro 'p12EpoRpJAK2'
 pEpoR_au = log10(offset_pEpoR_long + 16.0 * scale_pEpoR_long * (p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                                 ^~~~~~~~~~~~
82:123: error: expected ')' before ';' token
 pEpoR_au = log10(offset_pEpoR_long + 16.0 * scale_pEpoR_long * (p12EpoRpJAK2 + p1EpoRpJAK2 + p2EpoRpJAK2) / init_EpoRJAK2);
                                                                ~                                                          ^
                                                                                                                           )
In file included from nameless-mread-source.cpp:3:
nameless-mread-header.h:106:25: error: expected primary-expression before 'default'
 #define CIS (CIS_amt_ / default)
                         ^~~~~~~
83:34: note: in expansion of macro 'CIS'
 CIS_au = log10(offset_CIS_long + CIS * scale_CIS_long / CISEqc);
                                  ^~~
nameless-mread-header.h:106:25: error: expected ')' before 'default'
 #define CIS (CIS_amt_ / default)
             ~           ^~~~~~~
83:34: note: in expansion of macro 'CIS'
 CIS_au = log10(offset_CIS_long + CIS * scale_CIS_long / CISEqc);
                                  ^~~
[…]
