### 2021-11-01 #3708 @martinmodrak
Sorry to have kept everybody waiting. I finally managed to do a first pass through the software and paper.

This is a neat piece of software that looks well documented and well written. So far I haven't encountered any major obstacle.

My suggestions so far are minor, and may not even need immediate resolution for acceptance in JOSS.

1) The paper makes it look like Heta is the first tool with the ambition to build one modelling format that would serve many purposes. However, my (limited) understanding is that there are other contenders, notably SBML, which seems to have similar aims to Heta. I don't think I completely understand, what Heta adds that SMBL does not. If there are other

2) I filed a single bug (https://github.com/hetalang/heta-compiler/issues/10), which I presume is something relatively benign and easy to fix.

3) Possibly related is the fact that it appears (I didn't fully test) that the compiler is silent when an output format does not support a feature (e.g. not supporting some Switcher types). If that's the case, it might be helpful to raise a warning in such cases.

I still plan to spend some more time trying the software out, but wanted to share my results so far to let you react.
