### 2021-11-01 hetalang/heta-compiler#10 @metelkin
Hi @martinmodrak ,
Thank you for your feedback. I have the same issue.
The model included a compartment with "default" id.
This is a reserved word in mrgsolve, as I understand.

I've updated the repository with jak2stat5 model changing "default" => "def".
Now everything is fine. Compilation and simulations are OK.
Pull the latest commit from hetalang/qsp-benchmark and recompile the model.

This is a workaround. But the unexpected issues with reserved words are still a problem.
We need a better solution.
