### 2021-11-02 #3708 @metelkin
@martinmodrak,
Thank you for your review.
1) The Heta language is not the first formal representation of biomodels. There are formats of storage like SBML, CellML, Antimony, as well as some DSLs that are applied in a specific modeling tools like mrgsolve, ModelingToolkit.jl, etc. Each of them has some specific purpose and limitations. There were two reasons not to mention them in the article. The main reason was the limitation in manuscript size. A lot of different aspects are skipped because of it. The another reason is that the Heta modeling formats is the secondary topic. The main goal is to present Heta-compiler as a modeling environment and its useful features. Anyway I guess the question that you mentioned is important and I am going to do the following: (i) I will include one or two sentences regarding the other formats (I need to check how many symbols I have); (ii) I will write the extended comparison in documentation site.
2) I have fixed it and replied to the issue you mentioned. The problem was the reserved word "default" in mrgsolve. You are right. Some warning would be helpful. We will do it in future releases.
3) Actually Heta-compiler indicates errors and warnings if an export format does not support a feature or need attention. The Switcher component that you mentioned is the case. The compiler throws an error when exports to mrgsolve for unsupported features. But there are a lot of specific features (documented and undocumented) in each modeling format which we have to follow. I hope we will track them all in the future.
