### 2021-11-03 fd6137f
add toTerm() autotest
