### 2021-11-03 hetalang/heta-compiler#10 @martinmodrak
Wouldn't a systematic solution be that the `XXExport` classes expose a list of keywords and the compiler checks the model against this list. It could then either:

- Raise an error/warning when an id conflicts with the keywords, or
- Automatically rename all ids to avoid the conflict (e.g. by adding underscorses or something)

(I don't insist on this being done for the review, it just seems like a relatively cheap but useful solution)
