+++ src/core/unit-term.js
@@ -20,7 +20,7 @@ class UnitTerm extends Array {
     super();
     obj.length && obj.forEach((x) => { // filter((x) => x.exponent !== 0)
       if (termNames.indexOf(x.kind) === -1)
-        throw new TypeError('"kind" property of UnitTerm\'s item should one of reserved words, got ' + x.kind);
+        throw new TypeError('"kind" property of UnitTerm\'s item should be one of selected words, got ' + x.kind);
       this.push({
         kind: x.kind,
         exponent: (typeof x.exponent !== 'undefined') ? x.exponent : 1
