### 2021-11-04 hetalang/heta-compiler#10 @metelkin
@martinmodrak, agree, this would be good.
As a fast temporal solution, I've added "default" into a general keyword list.
Now the usage of such id throws an error. A user should change the id of such a component by himself.
But I still doubt if it's the best solution.
1) It seems that each "export" type should have its own stop list. We need time to collect and analyze them all.
2) We should also think of automatic renaming rules.

So I guess we suggest a better solution later.
