### 2021-11-04 hetalang/heta-compiler#11 @martinmodrak
CSwitcher is not handled by Mgrsolve, DBSolve, without a warning
I tried running the model under `cases/9-c-switcher`. I modified it to export also for MgrSolve. The exported SBML model (as run at http://core.fun.bio.keio.ac.jp/SBMLWebApp/) produces results that seem to take the switcher into account:

![image](https://user-images.githubusercontent.com/9483603/140330475-6275cb4a-902a-4982-9c9f-5943b43ea969.png)

However, the results in MgrSolve appear to ignore the switcher:

![image](https://user-images.githubusercontent.com/9483603/140330665-14836c1b-c9f1-4a3f-a16b-135d62fc3260.png)

The same result for DBSolve:

![image](https://user-images.githubusercontent.com/9483603/140330817-7842aa11-1358-4a21-90f7-110fd2293147.png)

However, there was no indication that CSwitcher is unsupported in either MgrSolve or DBSolve, here's the full output from the compilation:

```
Running compilation with declaration file "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\platform.json"...
[info] Builder initialized in directory "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher".
[info] Platform id: "cs"
[info] Compilation of module "src/index.heta" of type "heta"...
[info] Reading module of type "heta" from file "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\src\index.heta"...
[info] Setting references in elements, total length 51
[info] Checking for circular references in Records.
[warn] Units checking skipped. To turn it on set "unitsCheck: true" in declaration.
[info] Checking unit's terms.
[info] Start exporting to files, total: 7.
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\sbml.xml" of format "SBML"...
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\model" of format "SLV"...
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\model" of format "DBSolve"...
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\julia" of format "Julia"...
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\matlab" of format "Matlab"...
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\simbio" of format "Simbio"...
[warn] Cannot check units consistency for "r1" because no units found for "k1"
[warn] Cannot check units consistency for "r1" because no units found for "x1"
[warn] Cannot check units consistency for "r1" because no units found for "comp1"
[info] Exporting to "C:\Users\modrak_m\Documents\small\heta-review\heta-compiler-master\cases\9-c-switcher\dist\mrgsolve" of format "Mrgsolve"...
Compilation OK!
```
