### 2021-11-04 #3708 @martinmodrak
> The main goal is to present Heta-compiler as a modeling environment and its useful features. Anyway I guess the question that you mentioned is important and I am going to do the following: (i) I will include one or two sentences regarding the other formats (I need to check how many symbols I have); (ii) I will write the extended comparison in documentation site.

I agree the focus is elsewhere. I would just find it helpful to learn what I gain by writing the model in Heta and then compiling to SBML as opposed to directly writing SBML. I am choosing SBML as it appears to have quite a broad support, so presumably many workflows with Heta would involve compiling to SBML to get access to the vast ecosystem of tools that support it. One or two sentences would probably be needed to satisfy the "State of the field" requirement, however no need for an extended comparison just for me.

> Actually Heta-compiler indicates errors and warnings if an export format does not support a feature or need attention. The Switcher component that you mentioned is the case.

Unfortunately, this doesn't appear to be the case: https://github.com/hetalang/heta-compiler/issues/11

There is also potentially problematic design in the way the definition of export targets is part of the model specification. One could IMHO argue that there should be a separation between model definition and export targets. E.g. if I want to share my model with the community, I probably don't want to prescribe how the community will export it. Conflicting needs of different team members could lead to unnecessary version control conflicts in the export definitionsm etc. The need for `--julia-only` and `--skip-export` CLI options shows that you are struggling a bit with this design choice already.

Some (not mutually exclusive) options to consider:

- Move export definitions to `platform.json` (presumably, I won't be sharing `platform.json` the same way I share model files?, i.e. two team members could share Heta code but have each their own platform?)
- Allow full control of exports from the CLI (e.g. allow stuff like `heta build --export "{ filepath: model, format: DBSolve }"`)
- Keep the current system, but encourage separating export definitions and models in two files (probably by the export file including the model), codify this as a good practice (e.g. by using it in examples)
- Have a make-like system, where the export definitions function as "targets" and one can specify `heta build [target-list]` to only build specific targets)

Generally, handling exports well seems like an important (but hard) design decision. As with the other stuff, I don't insist on this being resolved for acceptance in JOSS, just wanted to share my concerns.

In summary, I think the only really blocking problem is fulfilling the "State of the field" requirement by at least briefly contrasting what one can do with Heta that one could not accomplish with SBML or other common tool. There are a bunch of suggestions and two bugs I noticed, but I am OK with acceptance even if those are not addressed.
