### 2021-11-05 7152fec
error messages for unsupported switchers
