+++ src/matlab-export/index.js
@@ -32,7 +32,7 @@ class MatlabExport extends AbstractExport {
     return 'MatlabExport';
   }
   get format(){
-    return 'Matlab'
+    return 'Matlab';
   }
   static get validate(){
     return ajv.compile(schema);
@@ -45,13 +45,13 @@ class MatlabExport extends AbstractExport {
     let logger = this._container.logger;
     if (this.spaceFilter.length === 0) {
       let msg = 'spaceFilter for Matlab format should include at least one namespace but get empty';
-      logger.err(msg);
+      logger.error(msg);
       var modelContent = '';
       var paramContent = '';
       var runContent = '';
     } else if (!this._container.namespaceStorage.has(this.spaceFilter[0])) {
       let msg = `Namespace "${this.spaceFilter[0]}" does not exist.`;
-      logger.err(msg);
+      logger.error(msg);
       modelContent = '';
       paramContent = '';
       runContent = '';
