### 2021-11-05 hetalang/heta-compiler#11 @metelkin
You are right, this is a bug. Heta-compiler must throw an error.
I've already pushed the commit to the master branch to fix it.

Currently, neither mrgsolve nor DBSovle supports conditional events at all.
There is no clear mechanism to implement such events into the formats.
It would be interesting to ask the mrgsolve developers what they think of implementing it.
