### 2021-11-05 #3708 @metelkin
> I would just find it helpful to learn what I gain by writing the model in Heta and then compiling to SBML as opposed to directly writing SBML. I am choosing SBML as it appears to have quite a broad support, so presumably many workflows with Heta would involve compiling to SBML to get access to the vast ecosystem of tools that support it. One or two sentences would probably be needed to satisfy the "State of the field" requirement, however no need for an extended comparison just for me.

I have added one sentence to mention the most known standard representation of bio-models. See the latest version in the "paper" branch. See below.

>The standardization of process-description modeling notation was also pursued in formats like SBML, CellML, Antimony. However the Heta standard can be distinguished by the specific features:  ... (list of Heta's features)

For a broader discussion here, I need to mention that SBML (as an example) and Heta are not competitors. Heta, in general, is a human-readable language that has been created to optimize the process of model development by users. SBML is an industrial xml-based standard for model exchange. I posted here a quote from SBML official site http://sbml.org
> SBML is a machine-readable format for representing models.
> What does SBML look like? Ugly. Don't look at it, unless you're developing software, in which case, you have to look at it, and we feel for you. SBML is really not meant to be edited by hand or exposed to users.
I know nobody who manually edits SBML in his everyday modeling workflow.

Anyway to develop the SBML model you have to use another layer: graphical programming (like in SimBiology), table creation (like in DBSolve), or use a DSL like in mrgsolve which is usually applicable for a single tool only.
