### 2021-11-05 #3708 @metelkin
> Unfortunately, this doesn't appear to be the case: hetalang/heta-compiler#11
Fixed in master branch. It will be in release v0.6.7
