### 2021-11-05 #3708 @metelkin
@martinmodrak, thanks again for such a deep analysis.
Regarding "export" design you are right. The current solution seems to be inconsistent in some way.
The "export" settings shouldn't be part of the standard and it is not part of a model at all.

Moreover, in one of the previous Heta versions "export" settings were part of the platfrom.json file.
The decision to move \#export back to the main Heta code was made based on multiple users requests.
It was easier for learning purposes.

I guess it is necessary to discuss the changes with endpoint users, i.e. modelers (some of them don't use GitHub at all).
It's problematic to modify this feature just now because many users use this option every day.

For now, I gonna create the issue and start this discussion there.
