### 2021-11-08 hetalang/heta-compiler#11 @martinmodrak
Great, looks like it works now. Looking at https://hetalang.github.io/#/heta-compiler/export-formats?id=features-support wouldn't a somewhat more general solution to the issue make sense? E.g. currently, the `@Dose` class is not handled by any export, but it appears no export checks whether the model includes it  and will thus likewise not produce an error.

Maybe it would make sense for each export to expose supported classes (via a new method) and then there could be a shared piece of code that checks that all classes used by the model are supported by the export? The biggest advantage would be that if you add a new class to the Heta standard, all exports you didn't touch would throw an error when this class is used, which seems like a sensible safe behavior.

But again this is a minor thing and I don't intend to pressure you into adopting my proposal.
