### 2021-11-08 #3708 @martinmodrak
Great. I think the software + paper now meets all criteria for acceptance in JOSS.
