### 2021-11-08 #3708 @fboehm
Thanks, everyone! Let me proofread the paper before recommending it for publication. I might have a small list of typos that need to be fixed. I expect to get back to you, @metelkin, in the next few days.
