### 2021-11-09 #3708 @fboehm
Line 20: Define SB before using it as an abbreviation. Does it abbreviate "Systems biology"?
Line 115: Capitalize the "H" in "Heta". One approach is to put curly braces around the H in the bib file: {H}eta
