### 2021-11-09 #3708 @fboehm
I can confirm that the DOIs in the bibliography all resolve to the intended articles.
