+++ paper/paper.bib
@@ -154,6 +154,6 @@ Heta compiler can be used as the framework for a QSP modeling project of any siz
    city = {Orlando, FL},
    journal = {American Conference on Pharmacometrics 10},
    month = {10},
-   title = {"Heta" is a New Declarative Language to Define the Large-Scale Systems Pharmacology and Systems Biology Models},
+   title = {"{H}eta" is a New Declarative Language to Define the Large-Scale Systems Pharmacology and Systems Biology Models},
    year = {2019},
 }
