+++ paper/paper.md
@@ -1,0 +1,104 @@
+---
+title: 'Heta compiler: a software tool for the development of large-scale QSP models and compilation into simulation formats'
+tags:
+  - systems biology
+  - quantitative systems pharmacology
+  - mathematical modeling
+  - pharmacometrics
+  - format conversion
+  - sbml
+  - heta
+authors:
+  - name: Evgeny Metelkin
+    orcid: 0000-0001-6612-5373
+    affiliation: 1
+affiliations:
+ - name: InSysBio LLC
+   index: 1
+date: 20 August 2021
+bibliography: paper.bib
+
+---
+
+# Summary
+
+Today mathematical modeling is becoming more and more popular in biomedicine and drug development. __Quantitative systems pharmacology__ (QSP), a relatively new research discipline, is devoted to complex models describing organisms, diseases, and drug dynamics. Designing these models presents a set of challenging methodological problems like managing a huge amount of data, dealing with large-scale models, time-consuming calculations, etc. __Heta compiler__ is a small and fast software tool written in JavaScript which manages infrastructure for QSP modeling projects. The purpose of the tool is to build and integrate QSP platform modules, to check their completeness and consistency, and then to compile everything into runnable code that will be executed in simulation software. A user can apply a command-line interface to run the model building process. Alternatively, Heta compiler can be used as a package for developing web-based applications or be integrated with simulation software.
+
+# Statement of need
+
+The large and still growing Systems Biology (SB) and Systems Pharmacology modeling communities utilize a variety of software tools for simulation and data analysis [@Stephanou2018; @Mentre2020; @Knight-Schrijver2016]. Usually, the modelers solve the algebraic-differential equations or perform parameters identification or sensitivity analysis. While being useful for tackling specific problems, each software tool often has no user-friendly way for routine operations like step-by-step model creation and maintenance. Furthermore, different tools have their own internal model format which cannot be reused.
+
+This paper presents Heta compiler which provides a convenient and flexible way for the development of dynamic large-scale models based on the __Heta language__ code. The compiler translates the source modeling code into a variety of formats to be run in simulation software tools. Heta compiler also provides information on errors in a model which can be used to debug.
+
+This tool is an effort to resolve the typical problems in a QSP project by creating a controllable working environment.
+The pre-formulated requirements are:  
+
+-	store QSP models and data in integrated infrastructure, 
+-	support iterative platform updates, 
+-	support of models written in human-readable formats as well as in tables, 
+-	help for model code reuse and sharing,
+- provide interface for storing several models in a single platform,
+-	export models and data to different popular formats so it can be used out-of-the-box.
+
+# Heta formats
+
+`Heta compiler` has been evolving alongside the Heta language [@metelkin2019] specification. Heta is a series of human-readable and writable formats for QSP and Systems Biology projects: Heta code, table representation, JSON, and YAML notation. Heta describes dynamic models in the process-description format i.e., as interacting components that describe volumes, concentrations, amounts, rates. On the other side, it was designed to be easily transformed into ODEs or other formats.
+
+The standardization of process-description modeling notation was also pursued in formats like SBML, CellML, Antimony. However the Heta standard can be distinguished by the specific features:  
+
+-	Human-readable/writable code that can be used for model development or modification.
+-	Easy code parsing.
+-	Modularity: QSP/SB platform can be subdivided into several files and spaces for better project management.
+- Multiple interchangeable representation: human-readable code, tables, JSON, YAML.
+-	Reusability: modeling platforms should be easily extended for other projects.
+-	Reach annotation capabilities for better model code revision.
[…]
