+++ paper/paper.bib
@@ -1,0 +1,159 @@
+@article{Smith2009,
+   author = {L. P. Smith and F. T. Bergmann and D. Chandran and H. M. Sauro},
+   doi = {10.1093/bioinformatics/btp401},
+   issn = {1367-4803},
+   issue = {18},
+   journal = {Bioinformatics},
+   month = {9},
+   title = {Antimony: a modular model definition language},
+   volume = {25},
+   year = {2009},
+}
+@article{Hucka2003,
+   author = {M. Hucka and A. Finney and H. M. Sauro and H. Bolouri and J. C. Doyle and H. Kitano and and the rest of the SBML Forum: and A. P. Arkin and B. J. Bornstein and D. Bray and A. Cornish-Bowden and A. A. Cuellar and S. Dronov and E. D. Gilles and M. Ginkel and V. Gor and I. I. Goryanin and W. J. Hedley and T. C. Hodgman and J.-H. Hofmeyr and P. J. Hunter and N. S. Juty and J. L. Kasberger and A. Kremling and U. Kummer and N. Le Novere and L. M. Loew and D. Lucio and P. Mendes and E. Minch and E. D. Mjolsness and Y. Nakayama and M. R. Nelson and P. F. Nielsen and T. Sakurada and J. C. Schaff and B. E. Shapiro and T. S. Shimizu and H. D. Spence and J. Stelling and K. Takahashi and M. Tomita and J. Wagner and J. Wang},
+   doi = {10.1093/bioinformatics/btg015},
+   issn = {1367-4803},
+   issue = {4},
+   journal = {Bioinformatics},
+   month = {3},
+   title = {The systems biology markup language (SBML): a medium for representation and exchange of biochemical network models},
+   volume = {19},
+   year = {2003},
+}
+@article{Kirouac2019,
+   author = {Daniel C. Kirouac and Brian Cicali and Stephan Schmidt},
+   doi = {10.1002/psp4.12390},
+   issn = {2163-8306},
+   issue = {4},
+   journal = {CPT: Pharmacometrics & Systems Pharmacology},
+   month = {4},
+   title = {Reproducibility of Quantitative Systems Pharmacology Models: Current Challenges and Future Opportunities},
+   volume = {8},
+   year = {2019},
+}
+@article{Mentre2020,
+   author = {France Mentré and Lena E. Friberg and Stephen Duffull and Jonathan French and Douglas A. Lauffenburger and Lang Li and Donald E. Mager and Vikram Sinha and Eric Sobie and Ping Zhao},
+   doi = {10.1002/cpt.1683},
+   issn = {0009-9236},
+   issue = {1},
+   journal = {Clinical Pharmacology & Therapeutics},
+   month = {1},
+   title = {Pharmacometrics and Systems Pharmacology 2030},
+   volume = {107},
+   year = {2020},
+}
+@article{Stephanou2018,
+   author = {Angélique Stéphanou and Eric Fanchon and Pasquale F. Innominato and Annabelle Ballesta},
+   doi = {10.1007/s10441-018-9330-2},
+   issn = {0001-5342},
+   issue = {4},
+   journal = {Acta Biotheoretica},
+   month = {12},
+   title = {Systems Biology, Systems Medicine, Systems Pharmacology: The What and The Why},
+   volume = {66},
+   year = {2018},
+}
+@article{Knight-Schrijver2016,
+   author = {V.R. Knight-Schrijver and V. Chelliah and L. Cucurull-Sanchez and N. Le Novère},
+   doi = {10.1016/j.csbj.2016.09.002},
+   issn = {20010370},
+   journal = {Computational and Structural Biotechnology Journal},
[…]
