+++ paper/README.md
@@ -1,0 +1,3 @@
+# JOSS paper
+
+This folder stores files for the publication in JOSS.
