### 2021-11-10 #3708 @metelkin
Hello @fboehm
I've published the release.

Version 0.6.7
10.5281/zenodo.5666487
