+++ paper/paper.bib
@@ -101,6 +101,7 @@ Heta compiler can be used as the framework for a QSP modeling project of any siz
 @article{Goryanin2006,
    author = {Igor I Goryanin and Galina V Lebedeva and Ekaterina A Mogilevskaya and Eugeniy A Metelkin and Oleg V Demin},
    issn = {0076-6941},
+   doi = {10.1002/0471973165.ch23},
    journal = {Methods of biochemical analysis},
    keywords = {Algorithms,Aminohydrolases,Aminohydrolases: metabolism,Bacterial,Bacterial Proteins,Bacterial Proteins: chemistry,Biological,Chemical,Escherichia coli,Escherichia coli: metabolism,Genes,Hydrogen-Ion Concentration,Kinetics,Models,Molecular Biology,Molecular Biology: methods,Systems Biology,Theoretical},
    month = {1},
