### 2021-11-10 #3708 @metelkin
@fboehm - I've added the required doi to the bib file.
